+++ changelog.md
@@ -28,6 +28,7 @@
 * Provide detailed errors when individual recipients fail
 * Report more errors when connecting
 * Add extras classes to composer classmap
+* Expose stream_context_create options via new SMTPOptions property
 
 ## Version 5.2.9 (Sept 25th 2014)
 * **Important: The autoloader is no longer autoloaded by the PHPMailer class**